changed gold to double

+++ src/gold.cc
@@ -3,11 +3,11 @@
 #include "visitor.h"
 using namespace std;
 
-Gold::Gold(int id, int value, bool lootable): Item(id,lootable,false), value(value){}
+Gold::Gold(int id, double value, bool lootable): Item(id,lootable,false), value(value){}
 
 Gold::~Gold(){}
 
-int Gold::getValue(){ return value; }
+double Gold::getValue(){ return value; }
 
 void Gold::setValue(int value) { this->value = value; }
